Refactor: Move spacingModifier file complete
